2022-06-14-database-mysql/main.py
import datetime
import mysql.connector

cnx = mysql.connector.connect(
    user='maicon',
    password='123456',
    host='127.0.0.1',
    database='employees')

query = ("SELECT first_name, last_name, hire_date FROM employees "
         "WHERE hire_date BETWEEN %s AND %s")

cursor = cnx.cursor()

hire_start = datetime.date(1999, 1, 1)
hire_end = datetime.date(1999, 12, 31)

cursor.execute(query, (hire_start, hire_end))

for (first_name, last_name, hire_date) in cursor:
  print("{}, {} was hired on {:%d %b %Y}".format(
    last_name, first_name, hire_date))

cursor.close()
cnx.close()
