2022-06-14-database-mysql/requirements.txt
autopep8==1.6.0
mysql-connector-python==8.0.29
protobuf==4.21.1
pycodestyle==2.8.0
toml==0.10.2
